server/requirements.txt
# Python dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
huggingface-hub==0.20.0
